version_elf+pe_release/ver.1_191107_ing/requirements.txt
numpy==1.16.4
pandas==0.24.2
pefile==2019.4.18
pyelftools==0.25
python-magic==0.4.15
PyYAML==5.1.2

